backend/requirements.txt
anyio==3.6.2
asyncio==3.4.3
asyncpg==0.27.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.98.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.1
Pillow==9.5.0
pydantic==1.10.9
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
